Re-organize the code a bit
Moved code around in routers so that the actual routes are
at the top and moved some functions to the utilities file.
