+++ python/src/etos_api/routers/v0/utilities.py
@@ -17,11 +17,17 @@
 import logging
 import asyncio
 import time
+
+import requests
+from fastapi import HTTPException
+from opentelemetry import trace
+
 from etos_api.library.graphql import GraphqlQueryHandler
 from etos_api.library.graphql_queries import (
     ARTIFACT_IDENTITY_QUERY,
     VERIFY_ARTIFACT_ID_EXISTS,
 )
+from etos_api.library.validator import SuiteValidator
 
 LOGGER = logging.getLogger(__name__)
 
@@ -67,3 +73,36 @@ async def wait_for_artifact_created(etos_library, artifact_identity, artifact_id
         await asyncio.sleep(2)
     LOGGER.error("Artifact %r not found.", artifact_identifier)
     return None
+
+
+async def download_suite(test_suite_url: str) -> dict:
+    """Attempt to download suite.
+
+    :param test_suite_url: URL to test suite to download.
+    :return: Downloaded test suite as JSON.
+    """
+    try:
+        suite = requests.get(test_suite_url, timeout=60)
+        suite.raise_for_status()
+    except Exception as exception:  # pylint:disable=broad-except
+        raise AssertionError(f"Unable to download suite from {test_suite_url}") from exception
+    return suite.json()
+
+
+async def validate_suite(test_suite_url: str) -> None:
+    """Validate the ETOS test suite through the SuiteValidator.
+
+    :param test_suite_url: The URL to the test suite to validate.
+    """
+    span = trace.get_current_span()
+
+    try:
+        test_suite = await download_suite(test_suite_url)
+        await SuiteValidator().validate(test_suite)
+    except AssertionError as exception:
+        LOGGER.error("Test suite validation failed!")
+        LOGGER.error(exception)
+        span.add_event("Test suite validation failed")
+        raise HTTPException(
+            status_code=400, detail=f"Test suite validation failed. {exception}"
+        ) from exception
